+++ clientele/generators/README.md
@@ -1,0 +1,7 @@
+# Generators
+
+In the future, this will be the directory for all the possible generators that clientele supports.
+
+## Standard
+
+The standard generator
