+++ clientele/generators/standard/templates/basic_client.jinja2
