+++ clientele/generators/standard/templates/bearer_client.jinja2
