+++ clientele/generators/standard/templates/client_py.jinja2
