+++ clientele/generators/standard/templates/config_py.jinja2
