+++ clientele/generators/standard/templates/get_method.jinja2
