+++ clientele/generators/standard/templates/http_py.jinja2
