+++ clientele/generators/standard/templates/manifest.jinja2
