+++ clientele/generators/standard/templates/post_method.jinja2
