+++ clientele/generators/standard/templates/schema_class.jinja2
