+++ clientele/generators/standard/templates/sync_methods.jinja2
