+++ clientele/generators/standard/writer.py
@@ -1,5 +1,11 @@
 from pathlib import Path
 
+from jinja2 import Environment, PackageLoader
+
+templates = Environment(
+    loader=PackageLoader("clientele", "generators/standard/templates/")
+)
+
 
 def write_to_schemas(content: str, output_dir: str) -> None:
     path = Path(output_dir) / "schemas.py"
